color names, truncate decimals
